+++ src/config_manager.py
@@ -3,6 +3,7 @@ from src.constants.config import CONFIG_SCHEME
 from src.constants.paths import CONFIG_PATH
 from src.sentinels import SENTINELS
 from src import gen_response
+from src.types import get_type_name
 from src.utils.misc import open_file
 
 class ConfigManager:
@@ -30,10 +31,21 @@ class ConfigManager:
                 SENTINELS.FROM_FILE: 'configure file'
             }[source]
 
+            type_ = get_type_name(CONFIG_SCHEME[name]['type'])
             default = CONFIG_SCHEME[name]['default']
             description = CONFIG_SCHEME[name]['description']
 
-            response = gen_response.Success(f'Got value of {name}', {'name': name, 'value': value, 'source': source, 'default': default, 'description': description})
+            response = gen_response.Success(
+                f'Got value of {name}', 
+                {
+                    'name': name, 
+                    'value': value, 
+                    'source': source,
+                    'type': type_,
+                    'default': default, 
+                    'description': description
+                    }
+                )
 
         return response