+++ src/constants/config_gui.py
@@ -1,0 +1,11 @@
+INIT_REMOTE = True
+TITLE = 'CADENCE Configure'
+RESIZE = (False, False)
+ICON_SIZE = (32, 32)
+
+FONT_SIZE = 15
+
+POP_UP_POS_X = 100 # relative to pos of master window
+POP_UP_POS_Y = 100
+
+DESC_WRAP_LEN = 500