+++ src/frontend/config_gui/__main__.py
@@ -1,0 +1,4 @@
+from .core import ConfigGUI
+
+if __name__ == '__main__':
+    ConfigGUI().run()